Restore testbed tests to passing state on Cocoa.
