+++ testbed/tests/window/test_window.py
@@ -1,11 +1,6 @@
 import gc
-import io
-import re
-import traceback
 import weakref
-from asyncio import wait_for
 from importlib import import_module
-from pathlib import Path
 from unittest.mock import Mock
 
 import pytest
@@ -26,6 +21,11 @@ async def second_window(second_window_kwargs):
 
 
 @pytest.fixture
+def default_window_title():
+    return "Toga"
+
+
+@pytest.fixture
 async def second_window_probe(app, app_probe, second_window):
     second_window.show()
     probe = window_probe(app, second_window)
@@ -35,20 +35,21 @@ async def second_window_probe(app, app_probe, second_window):
         second_window.close()
 
 
-async def test_title(main_window, main_window_probe):
-    """The title of a window can be changed"""
-    original_title = main_window.title
-    assert original_title == "Toga Testbed"
-    await main_window_probe.wait_for_window("Window title can be retrieved")
+@pytest.mark.parametrize("second_window_kwargs", [{}])
+async def test_title(second_window, second_window_probe):
+    """The title of a simple window can be changed"""
+    original_title = second_window.title
+    assert original_title == "Toga"
+    await second_window_probe.wait_for_window("Window title can be retrieved")
 
     try:
-        main_window.title = "A Different Title"
-        assert main_window.title == "A Different Title"
-        await main_window_probe.wait_for_window("Window title can be changed")
+        second_window.title = "A Different Title"
+        assert second_window.title == "A Different Title"
+        await second_window_probe.wait_for_window("Window title can be changed")
     finally:
-        main_window.title = original_title
-        assert main_window.title == "Toga Testbed"
-        await main_window_probe.wait_for_window("Window title can be reverted")
+        second_window.title = original_title
+        assert second_window.title == "Toga"
+        await second_window_probe.wait_for_window("Window title can be reverted")
 
 
 # Mobile platforms have different windowing characterics, so they have different tests.
@@ -56,18 +57,6 @@ if toga.platform.current_platform in {"iOS", "android"}:
     ####################################################################################
     # Mobile platform tests
     ####################################################################################
-
-    async def test_visibility(main_window, main_window_probe):
-        """Hide and close are no-ops on mobile"""
-        assert main_window.visible
-
-        main_window.hide()
-        await main_window_probe.wait_for_window("Window.hide is a no-op")
-        assert main_window.visible
-
-        main_window.close()
-        await main_window_probe.wait_for_window("Window.close is a no-op")
-        assert main_window.visible
 
     async def test_secondary_window():
         """A secondary window cannot be created"""
@@ -77,105 +66,27 @@ if toga.platform.current_platform in {"iOS", "android"}:
         ):
             toga.Window()
 
-    async def test_move_and_resize(main_window, main_window_probe, capsys):
-        """Move and resize are no-ops on mobile."""
-        initial_size = main_window.size
-        content_size = main_window_probe.content_size
-        assert initial_size[0] > 300
-        assert initial_size[1] > 500
-
-        assert main_window.position == (0, 0)
-
-        main_window.position = (150, 50)
-        await main_window_probe.wait_for_window("Main window can't be moved")
-        assert main_window.size == initial_size
-        assert main_window.position == (0, 0)
-
-        main_window.size = (200, 150)
-        await main_window_probe.wait_for_window("Main window cannot be resized")
-        assert main_window.size == initial_size
-        assert main_window.position == (0, 0)
-
-        try:
-            orig_content = main_window.content
-
-            box1 = toga.Box(
-                style=Pack(background_color=REBECCAPURPLE, width=10, height=10)
-            )
-            box2 = toga.Box(style=Pack(background_color=GOLDENROD, width=10, height=20))
-            main_window.content = toga.Box(
-                children=[box1, box2],
-                style=Pack(direction=COLUMN, background_color=CORNFLOWERBLUE),
-            )
-            await main_window_probe.wait_for_window("Main window content has been set")
-            assert main_window.size == initial_size
-            assert main_window_probe.content_size == content_size
-
-            # Alter the content width to exceed window width
-            box1.style.width = 1000
-            await main_window_probe.wait_for_window(
-                "Content is too wide for the window"
-            )
-            assert main_window.size == initial_size
-            assert main_window_probe.content_size == content_size
-
-            space_warning = (
-                r"Warning: Window content \([\d.]+, [\d.]+\) "
-                r"exceeds available space \([\d.]+, [\d.]+\)"
-            )
-            assert re.search(space_warning, capsys.readouterr().out)
-
-            # Resize content to fit
-            box1.style.width = 100
-            await main_window_probe.wait_for_window("Content fits in window")
-            assert main_window.size == initial_size
-            assert main_window_probe.content_size == content_size
-            assert not re.search(space_warning, capsys.readouterr().out)
-
-            # Alter the content width to exceed window height
-            box1.style.height = 2000
-            await main_window_probe.wait_for_window(
-                "Content is too tall for the window"
-            )
-            assert main_window.size == initial_size
-            assert main_window_probe.content_size == content_size
-            assert re.search(space_warning, capsys.readouterr().out)
-
-        finally:
-            main_window.content = orig_content
-
-    async def test_full_screen(main_window, main_window_probe):
-        """Window can be made full screen"""
-        main_window.full_screen = True
-        await main_window_probe.wait_for_window("Full screen is a no-op")
-
-        main_window.full_screen = False
-        await main_window_probe.wait_for_window("Full screen is a no-op")
-
-    async def test_screen(main_window, main_window_probe):
-        """The window can be relocated to another screen, using both absolute and relative screen positions."""
-        assert main_window.screen.origin == (0, 0)
-        initial_size = main_window.size
-        main_window.position = (150, 50)
-        await main_window_probe.wait_for_window("Main window can't be moved")
-        assert main_window.size == initial_size
-        assert main_window.position == (0, 0)
-        assert main_window.screen_position == (0, 0)
-
 else:
     ####################################################################################
     # Desktop platform tests
     ####################################################################################
 
     @pytest.mark.parametrize("second_window_kwargs", [{}])
-    async def test_secondary_window(app, second_window, second_window_probe):
+    async def test_secondary_window(
+        app, second_window, second_window_probe, default_window_title
+    ):
         """A secondary window can be created"""
         assert second_window.app == app
         assert second_window in app.windows
 
-        assert second_window.title == "Toga"
+        assert second_window.title == default_window_title
         assert second_window.size == (640, 480)
-        assert second_window.position == (100, 100)
+        # Position will be cascaded; the exact position depends on the platform,
+        # and how many windows have been created. As long as it's not (100,100),
+        # and the x an y coordinate are the same
+        assert second_window.position[0] > 110
+        assert second_window.position[1] == second_window.position[0]
+
         assert second_window_probe.is_resizable
         if second_window_probe.supports_closable:
             assert second_window_probe.is_closable
@@ -257,14 +168,9 @@ else:
         [dict(title="Secondary Window", position=(200, 300), size=(400, 200))],
     )
     async def test_secondary_window_toolbar(app, second_window, second_window_probe):
-        """A toolbar can be added to a secondary window"""
-        second_window.toolbar.add(app.cmd1)
-
-        # Window doesn't have content. This is intentional.
-        second_window.show()
-
-        assert second_window_probe.has_toolbar()
-        await second_window_probe.redraw("Secondary window has a toolbar")
+        """Simple windows don't have toolbars"""
+        with pytest.raises(AttributeError):
+            second_window.toolbar.add(app.cmd1)
 
     @pytest.mark.parametrize(
         "second_window_kwargs",
@@ -518,273 +424,3 @@ else:
                 second_window.position[0] - screen.origin[0],
                 second_window.position[1] - screen.origin[1],
             )
-
-
-async def test_as_image(main_window, main_window_probe):
-    """The window can be captured as a screenshot"""
-
-    screenshot = main_window.as_image()
-    main_window_probe.assert_image_size(
-        screenshot.size,
-        main_window_probe.content_size,
-        screen=main_window.screen,
-    )
-
-
-########################################################################################
-# Dialog tests
-########################################################################################
-
-
-TESTS_DIR = Path(__file__).parent
-
-
-async def assert_dialog_result(window, dialog, on_result, expected):
-    actual = await wait_for(dialog, timeout=1)
-    if callable(expected):
-        assert expected(actual)
-    else:
-        assert actual == expected
-
-    on_result.assert_called_once_with(window, actual)
-
-
-async def test_info_dialog(main_window, main_window_probe):
-    """An info dialog can be displayed and acknowledged."""
-    on_result_handler = Mock()
-    with pytest.warns(
-        DeprecationWarning,
-        match=r"Synchronous `on_result` handlers have been deprecated;",
-    ):
-        dialog_result = main_window.info_dialog(
-            "Info", "Some info", on_result=on_result_handler
-        )
-    await main_window_probe.redraw("Info dialog displayed")
-    await main_window_probe.close_info_dialog(dialog_result._impl)
-    await assert_dialog_result(main_window, dialog_result, on_result_handler, None)
-
-
-@pytest.mark.parametrize("result", [False, True])
-async def test_question_dialog(main_window, main_window_probe, result):
-    """An question dialog can be displayed and acknowledged."""
-    on_result_handler = Mock()
-    with pytest.warns(
-        DeprecationWarning,
-        match=r"Synchronous `on_result` handlers have been deprecated;",
-    ):
-        dialog_result = main_window.question_dialog(
-            "Question",
-            "Some question",
-            on_result=on_result_handler,
-        )
-    await main_window_probe.redraw("Question dialog displayed")
-    await main_window_probe.close_question_dialog(dialog_result._impl, result)
-    await assert_dialog_result(main_window, dialog_result, on_result_handler, result)
-
-
-@pytest.mark.parametrize("result", [False, True])
-async def test_confirm_dialog(main_window, main_window_probe, result):
-    """A confirmation dialog can be displayed and acknowledged."""
-    on_result_handler = Mock()
-    with pytest.warns(
-        DeprecationWarning,
-        match=r"Synchronous `on_result` handlers have been deprecated;",
-    ):
-        dialog_result = main_window.confirm_dialog(
-            "Confirm",
-            "Some confirmation",
-            on_result=on_result_handler,
-        )
-    await main_window_probe.redraw("Confirmation dialog displayed")
-    await main_window_probe.close_confirm_dialog(dialog_result._impl, result)
-    await assert_dialog_result(main_window, dialog_result, on_result_handler, result)
-
-
-async def test_error_dialog(main_window, main_window_probe):
-    """An error dialog can be displayed and acknowledged."""
-    on_result_handler = Mock()
-    with pytest.warns(
-        DeprecationWarning,
-        match=r"Synchronous `on_result` handlers have been deprecated;",
-    ):
-        dialog_result = main_window.error_dialog(
-            "Error", "Some error", on_result=on_result_handler
-        )
-    await main_window_probe.redraw("Error dialog displayed")
-    await main_window_probe.close_error_dialog(dialog_result._impl)
-    await assert_dialog_result(main_window, dialog_result, on_result_handler, None)
-
-
-@pytest.mark.parametrize("result", [None, False, True])
-async def test_stack_trace_dialog(main_window, main_window_probe, result):
-    """A confirmation dialog can be displayed and acknowledged."""
-    on_result_handler = Mock()
-    stack = io.StringIO()
-    traceback.print_stack(file=stack)
-    with pytest.warns(
-        DeprecationWarning,
-        match=r"Synchronous `on_result` handlers have been deprecated;",
-    ):
-        dialog_result = main_window.stack_trace_dialog(
-            "Stack Trace",
-            "Some stack trace",
-            stack.getvalue(),
-            retry=result is not None,
-            on_result=on_result_handler,
-        )
-    await main_window_probe.redraw(
-        f"Stack trace dialog (with{'out' if result is None else ''} retry) displayed"
-    )
-    await main_window_probe.close_stack_trace_dialog(dialog_result._impl, result)
-    await assert_dialog_result(main_window, dialog_result, on_result_handler, result)
-
-
-@pytest.mark.parametrize(
-    "filename, file_types, result",
-    [
-        ("/path/to/file.txt", None, Path("/path/to/file.txt")),
-        ("/path/to/file.txt", None, None),
-        ("/path/to/file.txt", ["txt", "doc"], Path("/path/to/file.txt")),
-        ("/path/to/file.txt", ["txt", "doc"], None),
-    ],
-)
-async def test_save_file_dialog(
-    main_window,
-    main_window_probe,
-    filename,
-    file_types,
-    result,
-):
-    """A file open dialog can be displayed and acknowledged."""
-    on_result_handler = Mock()
-    with pytest.warns(
-        DeprecationWarning,
-        match=r"Synchronous `on_result` handlers have been deprecated;",
-    ):
-        dialog_result = main_window.save_file_dialog(
-            "Save file",
-            suggested_filename=filename,
-            file_types=file_types,
-            on_result=on_result_handler,
-        )
-    await main_window_probe.redraw("Save File dialog displayed")
-    await main_window_probe.close_save_file_dialog(dialog_result._impl, result)
-
-    # The directory where the file dialog is opened can't be 100% predicted
-    # so we need to modify the check to only inspect the filename.
-    await assert_dialog_result(
-        main_window,
-        dialog_result,
-        on_result_handler,
-        None if result is None else (lambda actual: actual.name == result.name),
-    )
-
-
-@pytest.mark.parametrize(
-    "initial_directory, file_types, multiple_select, result",
-    [
-        # Successful single select
-        (TESTS_DIR, None, False, TESTS_DIR / "data.py"),
-        # Cancelled single select
-        (TESTS_DIR, None, False, None),
-        # Successful single select with no initial directory
-        (None, None, False, TESTS_DIR / "data.py"),
-        # Successful single select with file types
-        (TESTS_DIR, ["txt"], False, TESTS_DIR / "data.py"),
-        # Successful multiple selection
-        (
-            TESTS_DIR,
-            None,
-            True,
-            [TESTS_DIR / "conftest.py", TESTS_DIR / "data.py"],
-        ),
-        # Successful multiple selection of one item
-        (TESTS_DIR, None, True, [TESTS_DIR / "data.py"]),
-        # Cancelled multiple selection
-        (TESTS_DIR, None, True, None),
-        # Successful multiple selection with no initial directory
-        (None, None, True, [TESTS_DIR / "conftest.py", TESTS_DIR / "data.py"]),
-        # Successful multiple selection with file types
-        (
-            TESTS_DIR,
-            ["txt", "doc"],
-            True,
-            [TESTS_DIR / "conftest.py", TESTS_DIR / "data.py"],
-        ),
-    ],
-)
-async def test_open_file_dialog(
-    main_window,
-    main_window_probe,
-    initial_directory,
-    file_types,
-    multiple_select,
-    result,
-):
-    """A file open dialog can be displayed and acknowledged."""
-    on_result_handler = Mock()
-    with pytest.warns(
-        DeprecationWarning,
-        match=r"Synchronous `on_result` handlers have been deprecated;",
-    ):
-        dialog_result = main_window.open_file_dialog(
-            "Open file",
-            initial_directory=initial_directory,
-            file_types=file_types,
-            multiple_select=multiple_select,
-            on_result=on_result_handler,
-        )
-    await main_window_probe.redraw("Open File dialog displayed")
-    await main_window_probe.close_open_file_dialog(
-        dialog_result._impl, result, multiple_select
-    )
-    await assert_dialog_result(main_window, dialog_result, on_result_handler, result)
-
-
-@pytest.mark.parametrize(
-    "initial_directory, multiple_select, result",
-    [
-        # Successful single select
-        (TESTS_DIR, False, TESTS_DIR / "widgets"),
-        # Cancelled single select
-        (TESTS_DIR, False, None),
-        # Successful single select with no initial directory
-        (None, False, TESTS_DIR / "widgets"),
-        # Successful multiple selection
-        (TESTS_DIR, True, [TESTS_DIR, TESTS_DIR / "widgets"]),
-        # Successful multiple selection with one item
-        (TESTS_DIR, True, [TESTS_DIR / "widgets"]),
-        # Cancelled multiple selection
-        (TESTS_DIR, True, None),
-    ],
-)
-async def test_select_folder_dialog(
-    main_window,
-    main_window_probe,
-    initial_directory,
-    multiple_select,
-    result,
-):
-    """A folder selection dialog can be displayed and acknowledged."""
-    on_result_handler = Mock()
-    with pytest.warns(
-        DeprecationWarning,
-        match=r"Synchronous `on_result` handlers have been deprecated;",
-    ):
-        dialog_result = main_window.select_folder_dialog(
-            "Select folder",
-            initial_directory=initial_directory,
-            multiple_select=multiple_select,
-            on_result=on_result_handler,
-        )
-    await main_window_probe.redraw("Select Folder dialog displayed")
-    await main_window_probe.close_select_folder_dialog(
-        dialog_result._impl, result, multiple_select
-    )
-
-    if (
-        isinstance(result, list)
-        and not main_window_probe.supports_multiple_select_folder
-    ):
-        result = result[-1:]
-    await assert_dialog_result(main_window, dialog_result, on_result_handler, result)
